refactor(community): 상세 경로를 /community/post/{id}로 옮기고 댓글 영역을 목록·입력창으로 분리 (#271)
