+++ src/app/(main)/community/post/[postId]/page.tsx
